markets/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
pandas==2.1.4
redis==5.0.1
aioredis==2.0.1
httpx==0.25.2
pytz
pydantic==2.5.2
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0
